requirements.txt
faiss-cpu
numpy
streamlit
requests
langchain
sentence-transformers
crewai
ollama
langchain-community
pypdf
python-dotenv